Внёс правки преподавателя.
