+++ index.html
@@ -55,6 +55,6 @@
   </div>
 </body>
 
-<script type="module" src="./script.js"></script>
+<script type="module" src="./main.js"></script>
 
 </html>
