+++ modules/data.js
@@ -1,0 +1,2 @@
+export let comments = [];
+export let replyingTo = null;
