--- modules/tasks.js
@@ -1,11 +1,0 @@
-export let tasks = [
-    {
-        text: 'Купить чай',
-    },
-    {
-        text: 'Заварить чай',
-    },
-    {
-        text: 'Выпить чай',
-    },
-]
