+++ modules/utils.js
@@ -1,0 +1,17 @@
+export function escapeHtml(unsafe) {
+    return unsafe
+        .replace(/&/g, "&amp;")
+        .replace(/</g, "&lt;")
+        .replace(/>/g, "&gt;")
+        .replace(/"/g, "&quot;")
+        .replace(/'/g, "&#039;");
+}
+
+export function toggleLike(comments, commentId) {
+    const comment = comments.find(c => c.id === commentId);
+    if (comment) {
+        comment.isLiked = !comment.isLiked;
+        comment.likes += comment.isLiked ? 1 : -1;
+    }
+    return comments;
+}
